requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
black==24.4.0
blinker==1.7.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cx_Freeze==7.0.0
cx_Logging==3.2.0
Django==4.2.11
emoji==2.11.0
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
idna==3.8
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
lief==0.14.1
MarkupSafe==2.1.5
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
playsound==1.3.0
pycparser==2.22
pygame==2.5.2
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.29
sqlparse==0.4.4
tomli==2.0.1
twilio==9.2.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.1
yarl==1.9.7
zipp==3.18.1
